# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultima Casa API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/artworks": {
            "get": {
                "tags": [
                    "Artworks"
                ],
                "summary": "List all artworks",
                "description": "Returns paginated list of artworks",
                "operationId": "6888b4ce933c15ea45f2e8538dc6bb8d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Artwork"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Create a new artwork",
                "description": "Create a new artwork with images",
                "operationId": "96900a25c44caca8e817f91974e182d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ArtworkCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artwork created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artwork"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artworks/{id}": {
            "get": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Get artwork details",
                "description": "Returns detailed information about a specific artwork",
                "operationId": "f2cd78d2e8b7d5a584a78d6110d1a210",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artwork ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artwork"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artwork not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Update an artwork",
                "description": "Update artwork information",
                "operationId": "caf8f86c7210b6ddd61139879777689e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artwork ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ArtworkCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artwork updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artwork"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Artwork not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Delete an artwork",
                "description": "Delete a specific artwork",
                "operationId": "407924f3bd2fab834b2516a63e2a2143",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artwork ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artwork deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Artwork not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artworks/{artworkId}/images/{imageId}": {
            "delete": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Delete an artwork image",
                "description": "Delete a specific image from an artwork",
                "operationId": "0f9b8f8f61739dd88eacd7f2f9ecc316",
                "parameters": [
                    {
                        "name": "artworkId",
                        "in": "path",
                        "description": "Artwork ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "imageId",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Image deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Artwork or image not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Artwork": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "properties": {
                            "en": {
                                "type": "string"
                            },
                            "es": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "dimensions": {
                        "properties": {
                            "width": {
                                "type": "number"
                            },
                            "height": {
                                "type": "number"
                            },
                            "depth": {
                                "type": "number"
                            }
                        },
                        "type": "object"
                    },
                    "creation_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "artist": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArtworkImage"
                        }
                    }
                },
                "type": "object"
            },
            "ArtworkImage": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "is_primary": {
                        "type": "boolean"
                    },
                    "order": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ArtworkCreate": {
                "required": [
                    "title",
                    "images",
                    "creation_date"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "properties": {
                            "en": {
                                "type": "string"
                            },
                            "es": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    "dimensions": {
                        "properties": {
                            "width": {
                                "type": "number"
                            },
                            "height": {
                                "type": "number"
                            },
                            "depth": {
                                "type": "number"
                            }
                        },
                        "type": "object"
                    },
                    "creation_date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}